requirements.txt
pandas==2.1.0
scikit-learn==1.3.0
python-Levenshtein==0.21.1
fuzzywuzzy==0.18.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
joblib==1.3.2
python-dotenv==1.0.0
email-validator==2.0.0.post2
aiohttp==3.8.5
pytesseract==0.3.10
pdf2image==1.16.3
python-imap==1.0.0
beautifulsoup4==4.12.2
pymongo==4.5.0
pillow==10.0.0 